# ApiDocumentation/apiDocs.yml
openapi: 3.0.0
servers:
  - description: local Backend Address
    url: 'http://localhost:8080/'
info:
  description: >-
    Dies ist die API des Webshop Backends Die Enpunkte welche lediglich Basic
    Aut Calls Akzeptieren, enthalten nochmal eine gesonderte Beschreibung
    welcheRechte im Speziellen benötigt werden, dies ist im jeweiligen Enpunkt
    ersichtlich-
  version: 1.0.0
  title: WEBSHOP API
  contact:
    email: mirco_christoph.mueller@edu.fhdw.de
...
paths:
  /user/:
    get:
      security:
        - basicAuth: []
      tags:
        - user
      summary: login api returns user object
      operationId: login
      description: >
        by passing the correct user authentication via basic auth you get a json
        return value containing a json value.
      responses:
        '200':
          $ref: '#/components/responses/UserObject'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      security:
        - basicAuth: []
      tags:
        - user
      summary: updates userinformation and Role assignment
      description: >-
        this endpoint takes the Full or a partial Amount of the User Object, if
        a Role other than user is passed a validation check by the trequesting
        user is performed. if the requesting user is requesting a role assigment
        greater than it's own privileges e.g employee to admin then nothing will
        happen.
      responses:
        '200':
          $ref: '#/components/responses/UserObject'
        '400':
          description: something failed
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopUser'
        description: Updated user object
        required: true
    post:
      tags:
        - user
      summary: >-
        posts a user Object and stores it in Database could be seen as "register
        new user"
      responses:
        '200':
          $ref: '#/components/responses/UserObject'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopUser'
        description: Updated user object
        required: true
...
  /article/:
    get:
      summary: gets All Article Metadata Objects
      tags:
        - article
      responses:
        '200':
          $ref: '#/components/responses/ArticleMetadataList'
    post:
      summary: post a new Article Object
      tags:
        - article
      responses:
        '200':
          $ref: '#/components/responses/ArticleObject'
      requestBody:
        $ref: '#/components/requestBodies/UploadArticleWithPicture'
    put:
      summary: post an new Article Object with Picture
      tags:
        - article
      responses:
        '200':
          $ref: '#/components/responses/ArticleObject'
      requestBody:
        $ref: '#/components/requestBodies/UploadArticleWithPicture'
    delete:
      summary: deletes an article with Picture
      tags:
        - article
      responses:
        '200':
          description: success
          content:
            text/plain:
              example: true
...
  /artist/:
    get:
      summary: get all Artist Objects as json with Picture
      tags:
        - artist
      responses:
        '200':
          $ref: '#/components/responses/ArtistList'
    put:
      summary: 'changed an Article Object, based on the id in the json'
      security:
        - basicAuth: []
      tags:
        - artist
      parameters:
        - $ref: '#/components/responses/Artist'
      responses:
        '200':
          $ref: '#/components/responses/Artist'
    post:
      summary: adds a new Artist Object
      security:
        - basicAuth: []
      tags:
        - artist
      requestBody:
        $ref: '#/components/requestBodies/RegisterArtist'
      responses:
        '200':
          $ref: '#/components/responses/Artist'
    delete:
      summary: löscht einen Artikel
      security:
        - basicAuth: []
      tags:
        - artist
      parameters:
        - $ref: '#/components/responses/Artist'
      responses:
        '200':
          description: success
...
components:
  schemas:
    Article:
      type: object
      properties:
        artists:
          $ref: '#/components/schemas/Artist'
        description:
          type: string
        ean:
          format: int32
          type: integer
        genre:
          $ref: '#/components/schemas/Genre'
        id:
          format: int64
          type: integer
        image:
          $ref: '#/components/schemas/Picture'
        price:
          format: double
          type: number
        title:
          type: string
    ListArticle:
      type: array
      items:
        $ref: '#/components/schemas/Article'
    Artist:
      type: object
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        image:
    Genre:
      type: object
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
    Picture:
      type: object
      properties:
        article:
          $ref: '#/components/schemas/ListArticle'
        artist:
          $ref: '#/components/schemas/ListArtist'
        genre:
          $ref: '#/components/schemas/ListGenre'
        id:
          format: int64
          type: integer
    ListArtist:
      type: array
      items:
        $ref: '#/components/schemas/Artist'
    ListGenre:
      type: array
      items:
        $ref: '#/components/schemas/Genre'
...
    ShopUser:
      type: object
      properties:
        birth:
          $ref: '#/components/schemas/Date'
        email:
          pattern: \S
          type: string
          nullable: false
        firstName:
          type: string
        id:
          format: int64
          type: integer
        lastName:
          type: string
        password:
          type: string
        postalCode:
          format: int32
          type: integer
        role:
          $ref: '#/components/schemas/Role'
        street:
          type: string
        streetNumber:
          format: int32
          type: integer
        title:
          $ref: '#/components/schemas/Title'
        town:
          type: string
    Date:
      format: date
      type: string
    Role:
      enum:
        - ADMIN
        - EMPLOYEE
        - USER
      type: string
    Title:
      enum:
        - APACHE
        - BENUTZERDEFINIERT
        - DOKTOR
        - FRAU
        - HERR
        - PROFESSOR
      type: string
  requestBodies:
    UploadArticleWithPicture:
      description: >-
        Multipart Upload welcher den gleichzeitigen Upload mit einem Bild
        erlaubt
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              Article:
                $ref: '#/components/schemas/Article'
              fileName:
                type: string
                format: binary
          encoding:
            ArticlePicture:
              contentType: 'image/png, image/jpeg'
...
    RegisterArtist:
      description: erstellt einen neuen Artist Eintrag
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Artist'
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
    UserObject:
      description: returning json body with user object
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ShopUser'
...
    ArticleObject:
      description: returning a single Article Object
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              ArticleMetadata:
                $ref: '#/components/schemas/Article'
              ArticlePicture:
                $ref: '#/components/schemas/Picture'
          encoding:
            ArticlePicture:
              contentType: 'image/png, image/jpeg'
    ArticleMetadataList:
      description: Returns a List of All Metadata Objects
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Article'
    Artist:
      description: returns single Artist Object
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Artist'
...
    ArtistList:
      description: returns a List of ArtistObjects
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Artist'
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic